introduce test_exception.py
